@@ -1,6 +1,6 @@
 {
   "name": "eks-fargate-cluster",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "bin": {
     "eks-fargate-cluster": "bin/eks-fargate-cluster.js"
   },
@@ -10,18 +10,16 @@
     "cdk": "cdk"
   },
   "devDependencies": {
-    "@types/node": "10.17.5",
-    "aws-cdk": "1.95.1",
-    "ts-node": "^8.1.0",
-    "typescript": "^4.2.3"
+    "@aws-cdk/lambda-layer-kubectl-v25": "^2.0.3",
+    "@types/node": "18.15.11",
+    "aws-cdk": "^2.76.0",
+    "aws-cdk-lib": "^2.76.0",
+    "constructs": "^10.2.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.4"
   },
   "dependencies": {
-    "@aws-cdk/assert": "1.95.1",
-    "@aws-cdk/aws-ec2": "1.95.1",
-    "@aws-cdk/aws-eks": "1.95.1",
-    "@aws-cdk/aws-iam": "1.95.1",
-    "@aws-cdk/core": "1.95.1",
-    "@types/jest": "26.0.22",
-    "source-map-support": "^0.5.16"
+    "@types/jest": "29.5.1",
+    "source-map-support": "^0.5.21"
   }
 }
